[typed-collections] Move TypedPropertyFromToManyRelationTypeRector to TypedCollections namespace
