--- rules-tests/CodeQuality/Rector/Property/TypedPropertyFromToManyRelationTypeRector/Fixture/skip_empty_property.php.inc
@@ -1,8 +1,0 @@
-<?php
-
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Fixture;
-
-class SkipEmptyProperty
-{
-    public $training;
-}
