--- rules-tests/CodeQuality/Rector/Property/TypedPropertyFromToManyRelationTypeRector/Source/TrainingTerm.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Source;
-
-final class TrainingTerm
-{
-
-}
