+++ rules-tests/TypedCollections/Rector/Class_/ExplicitRelationCollectionRector/config/configured_rule.php
