+++ rules-tests/TypedCollections/Rector/Property/TypedPropertyFromToManyRelationTypeRector/Fixture/ODM/remove_default_null.php.inc
@@ -1,9 +1,8 @@
 <?php
 
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Fixture\ODM;
+namespace Rector\Doctrine\Tests\TypedCollections\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Fixture\ODM;
 
 use Doctrine\Common\Collections\ArrayCollection;
-use Rector\Doctrine\Tests\CodeQuality\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Source\TrainingTerm;
 
 final class RemoveDefaultNull
 {
@@ -25,13 +24,12 @@ final class RemoveDefaultNull
 namespace Rector\Doctrine\Tests\CodeQuality\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Fixture\ODM;
 
 use Doctrine\Common\Collections\ArrayCollection;
-use Rector\Doctrine\Tests\CodeQuality\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Source\TrainingTerm;
 
 final class RemoveDefaultNull
 {
     /**
      * @\Doctrine\ODM\MongoDB\Mapping\Annotations\EmbedMany(targetDocument="TrainingTerm")
-     * @var \Doctrine\Common\Collections\Collection<int, \Rector\Doctrine\Tests\CodeQuality\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Source\TrainingTerm>
+     * @var \Doctrine\Common\Collections\Collection<int, \Rector\Doctrine\Tests\TypedCollections\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Source\TrainingTerm>
      */
     private \Doctrine\Common\Collections\Collection $trainingTerms;
