+++ rules-tests/TypedCollections/Rector/Property/TypedPropertyFromToManyRelationTypeRector/Fixture/doctrine_attribute_relation_to_many.php.inc
@@ -1,20 +1,17 @@
 <?php
 
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Fixture;
+namespace Rector\Doctrine\Tests\TypedCollections\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Fixture;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
 
+#[ORM\Entity]
 class DoctrineRelationToMany
 {
-    /**
-     * @ORM\OneToMany(targetEntity="App\Product")
-     */
+    #[ORM\OneToMany(targetEntity:"App\Product")]
     private $products;
 
-    /**
-     * @ORM\ManyToMany(targetEntity="App\Car")
-     */
+    #[ORM\ManyToMany(targetEntity:"App\Car")]
     private $cars;
 
     public function __construct()
@@ -33,18 +30,19 @@ namespace Rector\Doctrine\Tests\CodeQuality\Rector\Property\TypedPropertyFromToM
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
 
+#[ORM\Entity]
 class DoctrineRelationToMany
 {
     /**
-     * @ORM\OneToMany(targetEntity="App\Product")
      * @var \Doctrine\Common\Collections\Collection<int, \App\Product>
      */
+    #[ORM\OneToMany(targetEntity:"App\Product")]
     private \Doctrine\Common\Collections\Collection $products;
 
     /**
-     * @ORM\ManyToMany(targetEntity="App\Car")
      * @var \Doctrine\Common\Collections\Collection<int, \App\Car>
      */
+    #[ORM\ManyToMany(targetEntity:"App\Car")]
     private \Doctrine\Common\Collections\Collection $cars;
 
     public function __construct()
