+++ rules-tests/TypedCollections/Rector/Property/TypedPropertyFromToManyRelationTypeRector/Fixture/skip_empty_property.php.inc
@@ -1,0 +1,8 @@
+<?php
+
+namespace Rector\Doctrine\Tests\TypedCollections\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Fixture;
+
+class SkipEmptyProperty
+{
+    public $training;
+}
