+++ rules-tests/TypedCollections/Rector/Property/TypedPropertyFromToManyRelationTypeRector/Source/TrainingTerm.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Doctrine\Tests\TypedCollections\Rector\Property\TypedPropertyFromToManyRelationTypeRector\Source;
+
+final class TrainingTerm
+{
+
+}
